Revert "feat: add LiquidGlassSegmentedStyle for segmented controls"
