+++ OffshoreBudgeting/Views/BudgetDetailsView.swift
@@ -344,7 +344,12 @@ private extension BudgetDetailsView {
                 .pickerStyle(.segmented)
                 .equalWidthSegments()
                 .frame(maxWidth: .infinity)
-                .liquidGlassSegmentedStyle()
+#if os(macOS)
+                .controlSize(.large)
+
+                .tint(themeManager.selectedTheme.glassPalette.accent)
+
+#endif
             }
             .padding(.horizontal, DS.Spacing.l)
             .ub_onChange(of: vm.selectedSegment) { newValue in
@@ -702,6 +707,8 @@ private struct FilterBar: View {
     let onChanged: () -> Void
     let onResetDate: () -> Void
 
+    @EnvironmentObject private var themeManager: ThemeManager
+
     var body: some View {
         GlassCapsuleContainer(
             horizontalPadding: DS.Spacing.l,
@@ -727,7 +734,12 @@ private struct FilterBar: View {
             }
             .pickerStyle(.segmented)
             .equalWidthSegments()
-            .liquidGlassSegmentedStyle()
+#if os(macOS)
+            .controlSize(.large)
+
+            .tint(themeManager.selectedTheme.glassPalette.accent)
+
+#endif
             .frame(maxWidth: .infinity)
         }
         .frame(maxWidth: .infinity)
